- make Tulip compiled with mingw64 without the -fpermissive flag (use correct semantic for casting a pointer to an integer type)
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@9679 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-core/src/PropertyTypes.cpp
@@ -20,6 +20,8 @@
 #include <tulip/PropertyTypes.h>
 #include <tulip/Graph.h>
 
+#include <stdint.h>
+
 using namespace std;
 using namespace tlp;
 
@@ -38,7 +40,7 @@ void GraphType::write(ostream &oss, const RealType &v) {
 }
 
 bool GraphType::read(istream& iss, RealType& v) {
-  unsigned long lv = 0;
+  uintptr_t lv = 0;
   bool ok = bool(iss >> lv);
 
   if (ok)
